+++ src/PhpSpreadsheet/Settings.php
@@ -21,11 +21,6 @@ class Settings
     private static ?string $chartRenderer = null;
 
     /**
-     * Default options for libxml loader.
-     */
-    private static ?int $libXmlLoaderOptions = null;
-
-    /**
      * The cache implementation to be used for cell collection.
      */
     private static ?CacheInterface $cache = null;
@@ -88,36 +83,6 @@ class Settings
     public static function htmlEntityFlags(): int
     {
         return ENT_COMPAT;
-    }
-
-    /**
-     * Set default options for libxml loader.
-     *
-     * @param ?int $options Default options for libxml loader
-     *
-     * @deprecated 3.5.0 no longer needed
-     */
-    public static function setLibXmlLoaderOptions(?int $options): int
-    {
-        if ($options === null) {
-            $options = defined('LIBXML_DTDLOAD') ? (LIBXML_DTDLOAD | LIBXML_DTDATTR) : 0;
-        }
-        self::$libXmlLoaderOptions = $options;
-
-        return $options;
-    }
-
-    /**
-     * Get default options for libxml loader.
-     * Defaults to LIBXML_DTDLOAD | LIBXML_DTDATTR when not set explicitly.
-     *
-     * @return int Default options for libxml loader
-     *
-     * @deprecated 3.5.0 no longer needed
-     */
-    public static function getLibXmlLoaderOptions(): int
-    {
-        return self::$libXmlLoaderOptions ?? (defined('LIBXML_DTDLOAD') ? (LIBXML_DTDLOAD | LIBXML_DTDATTR) : 0);
     }
 
     /**